Feat/backend api

+++ backend/db/models/Paper.js
@@ -1,11 +1,11 @@
-
 const { Model } = require('sequelize')
+
 module.exports = (sequelize, DataTypes) => {
   class Paper extends Model {}
   Paper.init(
     {
       id: {
-        type: DataTypes.BIGINT,
+        type: DataTypes.INTEGER,
         allowNull: false,
         primaryKey: true,
       },
